requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
sqlite3  # Built-in with Python

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1
python-bidi==0.4.2

# QR Code & Barcode
qrcode==7.4.2
python-barcode==0.15.1
pyzbar==0.1.9
Pillow==10.1.0

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.2
PyMuPDF==1.23.8

# Optional: zxing-cpp for better QR/barcode detection
# Uncomment to install (requires C++ compiler)
# zxing-cpp==2.2.0

# Excel Support
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dateutil==2.8.2



